Add XcodeUniversalSearch
I would like to add my new project, [XcodeUniversalSearch](https://github.com/sm11963/XcodeUniversalSearch), which I think is awesome! 😄 

I also noticed that Linex was a dead link and found what looks to be a clone of the original repo. Although, it is unclear what the maintenance level is since the App Store link is dead at that repo.